Add support for output hot-destroying amongst other changes
Not sure I like this direction I'm going, might back out of some of
these changes.
Also there's a fun bug where resizing the surface across two outputs
might cause endless configure events. I have no idea why that is and
I've been trying to figure it out for too long now :( will have another
look later..

+++ subprojects/nwl/seat.c
@@ -147,7 +147,7 @@ static const struct wl_keyboard_listener keyboard_listener = {
 };
 
 static void nwl_seat_set_pointer_cursor(struct nwl_seat *seat, struct nwl_surface *surface, const char *cursor) {
-	if (seat->state->cursor_theme_size != surface->scale*24) {
+	if ((int)seat->state->cursor_theme_size != surface->scale*24) {
 		if (seat->state->cursor_theme) {
 			wl_cursor_theme_destroy(seat->state->cursor_theme);
 		} else {
